refactor websocket rate limiter
